requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
ultralytics==8.2.0
pyttsx3==2.90
deepface==0.0.89
tf-keras==2.16.0
python-telegram-bot==21.3
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
Pillow==10.3.0
ollama==0.2.1